requirements.txt
# Core Libraries
python-dotenv
requests
pydantic
# FastAPI & Web Server
fastapi[standard]
uvicorn
# LLM & LangChain Ecosystem
langchain
langchain_community
langchain_text_splitters
langchain_huggingface
langchain-openai

# Document Handling

# Embedding & Vector Stores
sentence_transformers

# Databases
google.cloud
qdrant_client
pymongo
redis
bcrypt
